requirements.txt
gmpy2
pycryptodome
ecdsa